--- Intervot/src/shared/hooks/auth/login/useLoginForm.ts
@@ -1,25 +1,0 @@
-import { useForm } from "react-hook-form";
-
-type LoginForm = {
-  email: string;
-  password: string;
-};
-
-export const useLoginForm = () => {
-  const {
-    register,
-    handleSubmit,
-    formState: { errors, isValid },
-    trigger,
-  } = useForm<LoginForm>({
-    mode: "onChange",
-  });
-
-  return {
-    register,
-    handleSubmit,
-    errors,
-    isValid,
-    trigger,
-  };
-};
