+++ Intervot/src/shared/types/auth/login.ts
@@ -1,0 +1,4 @@
+export type LoginFormData = {
+  email: string;
+  password: string;
+};
