+++ Intervot/src/shared/types/auth/signup.ts
@@ -1,0 +1,12 @@
+export type SignupFormData = {
+  email: string;
+  nickname: string;
+  password: string;
+  confirmPassword: string;
+};
+
+export type SignupRequest = {
+  email: string;
+  nickname: string;
+  password: string;
+};
